+++ evio_util.c
@@ -19,8 +19,8 @@
 
 
 /* prototypes */
-static void parse_bank(unsigned long *buf, int ftype, int depth, NH_TYPE nh, LH_TYPE lh, void *userArg);
-static void loop_over_banks(unsigned long *data, int length, int type, int depth, NH_TYPE nh, LH_TYPE lh, void *userArg);
+static void parse_bank(unsigned int *buf, int ftype, int depth, NH_TYPE nh, LH_TYPE lh, void *userArg);
+static void loop_over_banks(unsigned int *data, int length, int type, int depth, NH_TYPE nh, LH_TYPE lh, void *userArg);
 
 
 /* container types used locally */
@@ -35,7 +35,7 @@ enum {
 /*---------------------------------------------------------------- */
 
 
-void evio_stream_parse(unsigned long *buf, NH_TYPE nh, LH_TYPE lh, void *userArg) {
+void evio_stream_parse(unsigned int *buf, NH_TYPE nh, LH_TYPE lh, void *userArg) {
 
   int depth=0;
   parse_bank(buf,BANK,depth,nh,lh,userArg);
@@ -47,7 +47,7 @@ void evio_stream_parse(unsigned long *buf, NH_TYPE nh, LH_TYPE lh, void *userArg
 /*---------------------------------------------------------------- */
 
 
-static void parse_bank(unsigned long *buf, int ftype, int depth, NH_TYPE nh, LH_TYPE lh, void *userArg) {
+static void parse_bank(unsigned int *buf, int ftype, int depth, NH_TYPE nh, LH_TYPE lh, void *userArg) {
 
   int length,tag,type,num,dataOffset;
 
@@ -135,7 +135,7 @@ static void parse_bank(unsigned long *buf, int ftype, int depth, NH_TYPE nh, LH_
 /*---------------------------------------------------------------- */
 
 
-static void loop_over_banks(unsigned long *data, int length, int type, int depth, NH_TYPE nh, LH_TYPE lh, void *userArg) {
+static void loop_over_banks(unsigned int *data, int length, int type, int depth, NH_TYPE nh, LH_TYPE lh, void *userArg) {
 
   int p=0;
